requirements.txt
entsoe-py
easyenergy
requests
energyzero>=5.0.0
nordpool
timezonefinder
reverse_geocoder
cryptography
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
aiohttp>=3.9.0
tenneteu-py>=0.1.0
pandas>=2.0.0
nednl>=0.2.0
holidays>=0.40
gie-py>=0.3.0
